requirements.txt
apache-airflow-providers-amazon
pandas
pyyaml
requests